refactor(lsp): split member completion helpers
